Index.js de la raiz refactorizado, optimizado y modularizado su código en diferentes archivos
